+++ tastyworks/models/option_chain.py
@@ -1,12 +1,16 @@
-from typing import Dict
-from dataclasses import dataclass
-
+import logging
 from datetime import date, datetime
-from tastyworks.models.option import Option,OptionType,OptionUnderlyingType
+from typing import Dict, List
+
+import aiohttp
+
+from tastyworks.models.option import Option, OptionType
+from tastyworks.models.underlying import Underlying, UnderlyingType
+
+LOGGER = logging.getLogger(__name__)
 
 
-@dataclass
-class OptionChain(object):   
+class OptionChain(object):
     """
     Maps option symbols to Option structures
     Provides filter methods based on type, strike price, expiry, etc.
@@ -15,7 +19,7 @@ class OptionChain(object):
         symbol = "AAPL"
         response = await client.get_option_chains(symbol)
         oc = OptionChain(OptionChain.parse(symbol, response))
-        
+
         for strike in oc.get_all_strikes():
             print(strike)
 
@@ -23,89 +27,61 @@ class OptionChain(object):
             print(expiry)
 
         for symbol, option in oc.get(strike=300.0,option_type=OptionType.CALL,expiry=datetime.date(2018,9,21)).items():
-            print(symbol, option) 
+            print(symbol, option)
     """
-    
-    chain : Dict[str, Option]
 
-    @classmethod
-    def parse(cls, ticker : str, response: dict):    
-        """
-        Parses a response from tasty session into a dict[str, Option]
+    def __init__(self, options):
+        self.options = options
 
-        Args:
-            ticker (str): Ticker symbol of the option chain
-            response (dict): Response from tasty session
+    def _get_filter_strategy(self, key, unique=True):
+        values = [getattr(option, key) for option in self.options]
+        if not any(values):
+            raise Exception(f'No values found for specified key: {key}')
 
-        Returns:
-            dict[str,Option]: dict mapping all option symbols to Option structures
-        """
+        values = list(set(values)) if unique else list(values)
+        return sorted(values)
 
-        chain = {}
-        for expiry, strikes in response.items():
-            for strike, opts in strikes.items():
-                chain[opts['call']] = Option(
-                    ticker=ticker,
-                    expiry=datetime.date(expiry),
-                    strike=strike,
-                    option_type=OptionType.CALL,
-                    underlying_type=OptionUnderlyingType.EQUITY, #todo pass underlying                                        
-                )
-
-                chain[opts['put']] = Option(
-                    ticker=ticker,
-                    expiry=datetime.date(expiry),
-                    strike=strike,
-                    option_type=OptionType.PUT,
-                    underlying_type=OptionUnderlyingType.EQUITY, #todo pass underlying                                        
-                )
-
-        
-        return chain
-
-    
-
-    
     def get_all_strikes(self):
-        return sorted(list(set([option.strike for symbol, option in self.chain.items()])))
+        return self._get_filter_strategy('strike')
 
     def get_all_expirations(self):
-        return sorted(list(set([option.expiry for symbol, option in self.chain.items()])))
-
-    def get(self, **kwargs):
-        """
-        Get filtered options
-
-        Args:
-            **kwargs (*): Option attribute and desired value
-        Returns:
-            dict: dict mapping option symbols to Option structure that pass the filter
-        """
-        
-        #Construct filter function
-        f = lambda o : all([ getattr(o, key) == val for key, val in kwargs.items()])
-
-        #Return filtered options
-        return {symbol : o for symbol, o in self.chain.items() if f(o)}       
-
-    def get_at_strike(self, strike):
-        return self.get(strike=strike)
-
-    def get_at_expiry(self, expiry):
-        return self.get(expiry=strike)
-
-    def get_all_puts(self):
-        return self.get(option_type=OptionType.PUT)
-    
-    def get_all_calls(self):
-        return self.get(option_type=OptionType.CALL)
-        
-
-    
+        return self._get_filter_strategy('expiry')
 
 
+async def get_option_chain(session, underlying: Underlying, expiration: date = None) -> List[Option]:
+    LOGGER.debug('Getting options chain for ticker: %s', underlying.ticker)
+    data = await _get_tasty_option_chain_data(session, underlying)
+    res = []
+
+    for exp in data['expirations']:
+        exp_date = datetime.strptime(exp['expiration-date'], '%Y-%m-%d').date()
+
+        if expiration and expiration != exp_date:
+            continue
+
+        for strike in exp['strikes']:
+            strike_val = float(strike['strike-price'])
+            for option_types in OptionType:
+                new_option = Option(
+                    ticker=underlying.ticker,
+                    expiry=exp_date,
+                    strike=strike_val,
+                    option_type=option_types,
+                    underlying_type=UnderlyingType.EQUITY
+                )
+                res.append(new_option)
+    return OptionChain(res)
 
 
+async def _get_tasty_option_chain_data(session, underlying) -> List[Dict]:
+    async with aiohttp.request(
+            'GET',
+            f'{session.API_url}/option-chains/{underlying.ticker}/nested',
+            headers=session.get_request_headers()) as response:
 
+        if response.status != 200:
+            raise Exception(f'Could not find option chain for symbol {underlying.ticker}')
+        resp = await response.json()
 
-
+        # NOTE: Have not seen an example with more than 1 item. No idea what that would be.
+        return resp['data']['items'][0]