+++ tastyworks/models/option_leg.py
@@ -1,5 +1,0 @@
-from tastyworks.models.model import Model
-
-
-class OptionLeg(Model):
-    pass