+++ tastyworks/models/security.py
@@ -1,6 +1,2 @@
-# NOTE: leaving this for now. Will continue when adding support for stocks
-from tastyworks.models.model import Model
-
-
-class Security(Model):
+class Security(object):
     pass