+++ tests/models/test_option.py
@@ -1,7 +1,8 @@
 import unittest
 from datetime import date
 
-from tastyworks.models.option import Option, OptionType, OptionUnderlyingType
+from tastyworks.models.option import Option, OptionType
+from tastyworks.models.underlying import UnderlyingType
 
 
 class TestOptionModel(unittest.TestCase):
@@ -12,7 +13,7 @@ class TestOptionModel(unittest.TestCase):
             expiry=date(2018, 8, 10),
             strike=3.5,
             option_type=OptionType.CALL,
-            underlying_type=OptionUnderlyingType.EQUITY
+            underlying_type=UnderlyingType.EQUITY
         )
 
     def test_occ2010_integer_strike(self):
@@ -54,3 +55,16 @@ class TestOptionModel(unittest.TestCase):
 
         res = self.test_option.get_occ2010_symbol()
         self.assertEqual(expected_result, res)
+
+    def test_get_dxfeed_symbol(self):
+        expected_result = 'AKS180810C3.5'
+        result = self.test_option.get_dxfeed_symbol()
+        self.assertEqual(result, expected_result)
+
+    def test_no_option_chains_on_option(self):
+        with self.assertRaises(Exception):
+            self.test_option.get_option_chain()
+
+    def test_get_underlying_type_string(self):
+        res = self.test_option._get_underlying_type_string(UnderlyingType.EQUITY)
+        self.assertEqual(res, 'Equity Option')